bugfix(android): Allow `Volume*` keys to be passed to the user
It seems that discussions around #2748 being a poor workaround that
should rather be implemented properly/differently drowned out the fact
that the workaround was implemented incorrectly to begin with.  By
having the `if` inside the match arm, rather than /on/ the match arm
the `keycode =>` path is /never/ used for `Volume*` keys and this
event is thus never passed on to the user, defeating the purpose
of calling `handle_volume_keys()`.  Only the `Handled`/`Unhandled`
state is affected by the `ignore_volume_keys` workaround, as set by
`EventLoopBuilderExtAndroid::handle_volume_keys()`.

Fix this by moving the `if` workaround on the match arm, so that
disabling it via `EventLoopBuilderExtAndroid::handle_volume_keys()`
causes `Volume*` events to be delivered to user code once again.